Variational inference of (factored) regression.

Variational inference of (factored) regression.

fit.fqtl(y, x.mean, factored = FALSE, svd.init = TRUE,
  model = c("gaussian", "nb", "logit", "voom", "beta"), c.mean = NULL,
  x.var = NULL, y.loc = NULL, y.loc2 = NULL, x.mean.loc = NULL,
  c.mean.loc = NULL, cis.dist = 1e+06, weight.nk = NULL,
  do.hyper = FALSE, tau = NULL, pi = NULL, tau.lb = -10,
  tau.ub = -4, pi.lb = -4, pi.ub = -1, tol = 1e-04,
  gammax = 1000, rate = 0.01, decay = 0, jitter = 0.1,
  nsample = 10, vbiter = 2000, verbose = TRUE, k = 1,
  right.nn = FALSE, mu.min = 0.01, print.interv = 10, nthread = 1,
  rseed = NULL, options = list())

Arguments

y
[n x m] response matrix

x.mean
[n x p] primary covariate matrix for mean change (can specify location)

factored
is it factored regression? (default: FALSE)

svd.init
Initalize by SVD (default: TRUE)

model
choose an appropriate distribution for the generative model of y matrix from c('gaussian', 'nb', 'logit', 'voom', 'beta') (default: 'gaussian')

c.mean
[n x q] secondary covariate matrix for mean change (dense)

x.var
[n x r] covariate marix for variance#'

y.loc
m x 1 genomic location of y variables

y.loc2
m x 1 genomic location of y variables (secondary)

x.mean.loc
p x 1 genomic location of x.mean variables

c.mean.loc
q x 1 genomic location of c.mean variables

cis.dist
distance cutoff between x and y

weight.nk
(non-negative) weight matrix to help factors being mode interpretable

do.hyper
Hyper parameter tuning (default: FALSE)

tau
Fixed value of tau

pi
Fixed value of pi

tau.lb
Lower-bound of tau (default: -10)

tau.ub
Upper-bound of tau (default: -4)

pi.lb
Lower-bound of pi (default: -4)

pi.ub
Upper-bound of pi (default: -1)

tol
Convergence criterion (default: 1e-4)

gammax
Maximum precision (default: 1000)

rate
Update rate (default: 1e-2)

decay
Update rate decay (default: 0)

jitter
SD of random jitter for mediation & factorization (default: 0.01)

nsample
Number of stochastic samples (default: 10)

vbiter
Number of variational Bayes iterations (default: 2000)

verbose
Verbosity (default: TRUE)

k
Rank of the factored model (default: 1)

right.nn
non-negativity on the right side of the factored effect (default: FALSE)

mu.min
mininum non-negativity weight (default: 0.01)

print.interv
Printing interval (default: 10)

nthread
Number of threads during calculation (default: 1)

rseed
Random seed

options
A combined list of inference/optimization options

Value

a list of variational inference results

Details

Estimate factored or non-factored SNP x gene / tissue multivariate association matrix. More precisely, we model mean parameters of the Gaussian distribution by either factored mean

$$\mathsf{E}[Y] = X \theta_{\mathsf{snp}} \theta_{\mathsf{gene}}^{\top} + C \theta_{\mathsf{cov}}$$

or independent mean $$\mathsf{E}[Y] = X \theta + C \theta_{\mathsf{cov}}$$

and variance $$\mathsf{V}[Y] = X_{\mathsf{var}} \theta_{\mathsf{var}}$$

Each element of mean coefficient matrix follows spike-slab prior; variance coefficients follow Gaussian distribution.

Examples